Miscellaneous bugfixes for the Blockly update
This is a squashed commit containing bugfixes for the Blockly
update. The individual commit bodies are below as a summary:

1. After using the Paste All Blocks from Backpack menu item, undoing
this operation requires N Undo commands whereas it should only require
1 Undo. This commit wraps the bulk of the pasteBackpack operation in a
try-finally to set/unset the Blockly event group so that all of the
create block events are grouped together.

2. Prevent CSS updates during paste all from backpack

3. Ghica reported that when the backpack is open it is possible to
scroll both the flyout's and workspace's scrollbars and there is no
visual indicator which is which. This means that sometimes the user
will scroll the workspace when intending to scroll the flyout. This
commit makes flyouts hide their parent workspace's scrollbars and show
them after the flyout is dismissed.

4. Improve rendering of flyouts with many blocks

5. Hide backpack with ESC key

6. Only fire parameter orientation event after orientation change

7. The implementation of Blockly.Events.filer on Google's master is
implemented using a O(n^2) algorithm. When performing operations that
affect a large number of blocks, e.g. arrangements or pasting from the
Backpack, this performance is unacceptable. This commit provides a new
implementation of Blockly.Events.filter that is linear in the number
of pending events.

8. Removing mutators for parameters or deleting the procedure block
and then attempting to undo the operation would leave the caller
blocks in an inconsistent state with missing parameters or not
reconnecting the previously attached input blocks. This commit fixes
these issues by changing the order of certain operations so that when
Blockly reverses the order of operations variables, such as the caller
block's PROCNAME, are set correctly.

9. The way that Blockly handles changes to fields breaks how we
renamed parameters in the mutator. This commit provides a custom
handler for key events that makes the renaming behavior match that
before the Blockly update.

10. Cherry-picked code for mutators that fixed the ordering of the
mutator workspace's toolbox with respect to the block canvas.

11. Fixed delete area not appearing for the mutator workspace.

12. Fixed ghost scrollbars from different flyouts capturing the mouse
over event for blocks in the drawer flyout.

13. Made it so that dragging a block in any direction from a flyout
will result in a new block being created. Previously the window was
±70 degrees from the horizontal axis.

14. Improved performance on large projects during load or after moving
a stack of blocks.

15. Fixed error checking on blocks created by typeblock and their
relatives.

16. Fixed null pointer exception when renaming a component prior to
opening the blocks editor.

17. Moved updates to the scrollbar visibility to setScrollbarsVisible
and made all changes to scrollbar visibility go through that method,
which includes fix 16 above.

18. Made arrangement options adjust to workspace scale and
snap-to-grid options.

19. Fixed issue where changing text fields would not result in a
Change event and therefore fail to trigger a workspace save.

20. Added fix to ensure the warning handler exists before attempting
to do error checks on variables.

21. Patched Blockly code to prevent the connection database from
throwing an error when an operation would essentially be an identity
operation on the database.

+++ appinventor/blocklyeditor/src/blocks/procedures.js
@@ -196,10 +196,8 @@ Blockly.Blocks['procedures_defnoreturn'] = {
       }
       this.render();
     }
-    if (Blockly.Events.isEnabled()) {
-      // Trigger a Blockly UI change event
-      Blockly.Events.fire(new Blockly.Events.Ui(this, 'parameter_orientation',
-        (!this.horizontalParameters).toString(), this.horizontalParameters.toString()))
+    if (this.workspace.loadCompleted) {  // set in BlocklyPanel.java on successful load
+      Blockly.Procedures.mutateCallers(this);
     }
     // console.log("exit procedures_defnoreturn updateParams_()");
   },
@@ -297,6 +295,11 @@ Blockly.Blocks['procedures_defnoreturn'] = {
     if (params.length != 0 && isHorizontal !== this.horizontalParameters) {
       this.horizontalParameters = isHorizontal;
       this.updateParams_();
+      if (Blockly.Events.isEnabled()) {
+        // Trigger a Blockly UI change event
+        Blockly.Events.fire(new Blockly.Events.Ui(this, 'parameter_orientation',
+          (!this.horizontalParameters).toString(), this.horizontalParameters.toString()));
+      }
     }
   },
   mutationToDom: function() {
@@ -375,17 +378,20 @@ Blockly.Blocks['procedures_defnoreturn'] = {
     var editable = this.editable_;
     var workspace = this.workspace;
 
-    // Call parent's destructor.
-    Blockly.BlockSvg.prototype.dispose.apply(this, arguments);
-
+    // This needs to happen first so that the Blockly events will be replayed in the correct
+    // order on undo
     if (editable) {
       // Dispose of any callers.
       //Blockly.Procedures.disposeCallers(name, workspace);
       Blockly.AIProcedure.removeProcedureValues(name, workspace);
-      var procDb = workspace.getProcedureDatabase();
-      if (procDb) {  // only true for the top-level workspaces, not flyouts/flydowns
-        procDb.removeProcedure(this.id);
-      }
+    }
+
+    // Call parent's destructor.
+    Blockly.BlockSvg.prototype.dispose.apply(this, arguments);
+
+    var procDb = workspace.getProcedureDatabase();
+    if (editable && procDb) {  // only true for the top-level workspaces, not flyouts/flydowns
+      procDb.removeProcedure(this.id);
     }
 
   },
@@ -557,9 +563,21 @@ Blockly.Blocks['procedures_mutatorarg'] = {
 //      return Blockly.LexicalVariable.renameParam.call(this,newName);
 //    }
     this.setColour(Blockly.PROCEDURE_CATEGORY_HUE);
+    var editor = new Blockly.FieldTextInput('x',Blockly.LexicalVariable.renameParam);
+    // 2017 Blockly's text input change breaks our renaming behavior.
+    // The following is a version we've defined.
+    editor.onHtmlInputChange_ = function(e) {
+      var oldValue = this.getValue();
+      Blockly.FieldFlydown.prototype.onHtmlInputChange_.call(this, e);
+      var newValue = this.getValue();
+      if (newValue && oldValue !== newValue && Blockly.Events.isEnabled()) {
+        Blockly.Events.fire(new Blockly.Events.Change(this.sourceBlock_, 'field', this.name,
+          oldValue, newValue));
+      }
+    };
     this.appendDummyInput()
         .appendField(Blockly.Msg.LANG_PROCEDURES_MUTATORARG_TITLE)
-        .appendField(new Blockly.FieldTextInput('x',Blockly.LexicalVariable.renameParam), 'NAME');
+        .appendField(editor, 'NAME');
     this.setPreviousStatement(true);
     this.setNextStatement(true);
     this.setTooltip(Blockly.Msg.LANG_PROCEDURES_MUTATORARG_TOOLTIP);
@@ -830,10 +848,11 @@ Blockly.Blocks['procedures_callnoreturn'] = {
     options.push(option);
   },
   removeProcedureValue: function() {
-    this.setFieldValue("none", 'PROCNAME');
+    // Detach inputs before resetting name so that undo/redo operations happen in the right order
     for(var i=0;this.getInput('ARG' + i) !== null;i++) {
       this.removeInput('ARG' + i);
     }
+    this.setFieldValue("none", 'PROCNAME');
   },
   // This generates a single generic call to 'call no return' defaulting its value
   // to the first procedure in the list. Calls for each procedure cannot be done here because the
